converted the functionality into a bootstrap question
